+++ src/pages/NotificationCenter/PriceAlerts/ActiveAlerts/index.tsx
@@ -1,19 +1,17 @@
-import { Trans } from '@lingui/macro'
+import { t } from '@lingui/macro'
 import { useEffect, useState } from 'react'
-import { Info } from 'react-feather'
-import { Flex, Text } from 'rebass'
+import { Flex } from 'rebass'
 import { useGetAlertStatsQuery, useGetListAlertsQuery } from 'services/priceAlert'
 
 import Loader from 'components/Loader'
 import { useActiveWeb3React } from 'hooks'
-import useTheme from 'hooks/useTheme'
+import NoData from 'pages/NotificationCenter/NoData'
 import CommonPagination from 'pages/NotificationCenter/PriceAlerts/CommonPagination'
 import { ITEMS_PER_PAGE } from 'pages/NotificationCenter/const'
 
 import SingleAlert from './SingleAlert'
 
 const ActiveAlerts = ({ setDisabledClearAll }: { setDisabledClearAll: (v: boolean) => void }) => {
-  const theme = useTheme()
   const { account } = useActiveWeb3React()
   const [page, setPage] = useState(1)
   const { data, isLoading } = useGetListAlertsQuery(
@@ -41,23 +39,7 @@ const ActiveAlerts = ({ setDisabledClearAll }: { setDisabledClearAll: (v: boolea
   }
 
   if (!data?.alerts?.length) {
-    return (
-      <Flex flex="1 1 0" justifyContent="center" width="100%" alignItems="center">
-        <Flex
-          sx={{
-            flexDirection: 'column',
-            alignItems: 'center',
-            color: theme.subText,
-            gap: '0.75rem',
-          }}
-        >
-          <Info size={'24px'} />
-          <Text as="span">
-            <Trans>No price alerts created yet</Trans>
-          </Text>
-        </Flex>
-      </Flex>
-    )
+    return <NoData msg={t`No price alerts created yet`} />
   }
 
   return (
